Added questions and their answers Added questions class and Question array which both shows the prompts(the actual questions) and their answers(from this.answer and this.prompt as well(
Tomorrow or in the upcoming days, i will add a for loop which would display the questions and also add up (based on the 11 questions) how much points they earned.

Depending on this outcomes, if the player scores either a 7 or 8 out of 11, they are consisdered fellow zoes, if not, yo pas zoe's. Yo pas haitiens(ou yo sont mais yo vleh pas dit ça.